refactor: move constants to hooks directory and update imports
