+++ src/components/Page/Archive/Pagination/PostList.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react';
 import LinkCard from '@/components/UI/LinkCard';
 import { convertPostSlugToPath } from '@/lib/url';
 import type { TermsPostListProps } from '@/types/source';
-import { ITEMS_PER_PAGE, QUERY_PAGE_KEY } from './constant';
+import { ITEMS_PER_PAGE, QUERY_PAGE_KEY } from './hooks/constant';
 
 type PostListProps = {
   posts: TermsPostListProps[];
