+++ src/components/Page/Archive/Pagination/index.ts
@@ -1,5 +1,5 @@
 export { PaginationComponent } from './components/PaginationComponent';
-export { DOTS, ITEMS_PER_PAGE, QUERY_PAGE_KEY } from './constant';
+export { DOTS, ITEMS_PER_PAGE, QUERY_PAGE_KEY } from './hooks/constant';
 export { usePagination } from './hooks/usePagination';
 export { Pagination } from './Pagination';
 export { PostList } from './PostList';
